Fix: overwriting stairs in movement tests
Like some of the other coordinates in testmove.lua (addressed in
20ef4c6), the attempt to overwrite the stairs was targeting a spot
one space to the left of the actual position of the stairs, causing it
to have no effect (discussion suggested this may have been a result of
ff46d96).  Update it so the stairs are properly erased and won't
interfere with movement tests by stopping a running/rushing hero early.