menu_drop() vs ECMD_TIME
Some routines return ECMD_TIME|ECMD_CANCEL (for instance when 'a'pply
wields an item and player cancels the attempt to use it) so change
drop_menu() to test that properly.  I don't think drop() ever returns
that combined mask value but be prepared to handle time passage if it
ever does.